Move FlowFactor and _base_model_metadata into core
FlowFactor is domain-agnostic -- it reaches data only through the
SamplerContext.prepared_data() protocol -- so it belongs in core alongside the
Factor ABC rather than in the GW module. The GW layer keeps only the GW-specific
pieces: GWSamplerContext (which builds the data-prep conditioning map f_i) and
GWComposedSampler (post-processing). Document GWSamplerContext.raw_context
(= EventDataset.data; retained for the not-yet-wired likelihood and the
serialized transport state).

dingo.gw.inference.factors re-exports FlowFactor, so existing imports keep
working. Plain-NPE parity with GWSampler remains bit-exact.
